transcoder_service/pyproject.toml
[project]
name = "transcoder-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "boto3>=1.39.1",
    "ffmpeg-python>=0.2.0",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy[asyncio]>=2.0.41",
]
